refactor*: rsvp: Clean-up terminology around ids
There are two ways to identify a calendar event:

- by globally-unique event uid,
- by Proton-specific event id.

Since the external systems that generate invites don't really care about
Proton-specific identifiers, the `invite.ics` we get will mostly contain
only the event uid (next to its other metadata, of course).

UIDs are fine and we can map them into Proton-specific ids when needed:

https://gitlab.protontech.ch/rust/proton-rust/-/blob/77bc0d9dfd1f60a7e03709c48d1f062e194549fb/calendar/calendar-common/src/rsvp/fetch.rs#L53

... but there are cases where we *do* actually get the Proton-specific
id directly from email headers - those are:

- invites from external system (backend auto-imports them and adds an
  extra "fyi" header),
- e-mail reminders.

This prompted me to design `RsvpEventId` as:

```rust
enum RsvpEventId {
    // Event for which we know the Proton-specific id:
    Direct(CalendarEventId, ...),

    // Event for which we need to look-up the Proton-specific id:
    Indirect(CalendarEventUid, ...),
}
```

... where `RsvpEventId::Direct` means we've basically short-circuited
the code by loading event id from headers instead of analyzing the
`invite.ics` file (which would then yield `RsvpEventId::Indirect`).

But this is actually wrong - the top-level structure must be:

```rust
enum RsvpEventId {
    Invite(CalendarEventUid, ...),
    Reminder(CalendarEventId, ...),
}
```

... because even if backend provided us with the Proton-specific event
id, we still need to load `invite.ics` to make sure the invite did not
get out of date.

Example scenario, as an organizer:

- create an event (this sends e-mail numero uno),
- update event's start/end date (this sends e-mail numero due).

When an attendee opens e-mail numero uno, we need to be able to tell
that this is a stale invite - the way we do that is by comparing the
`DTSTAMP` property from the `invite.ics` to the `DTSTAMP` returned by
the API; if our attachment's dtstamp is in the past, the invite is
stale and the user should be informed.

Or rather that's what we're supposed to do eventually - that's what I
started to implement in a separate branch today before I realized that
for external invites we always end up with `RsvpEventId::Direct` (i.e.
the event id loaded from headers), so there's no way to figure out if
the invite was stale!

tl;dr we need to load `invite.ics` always, for both internal and
external invites, and the only case where we can utilize mail headers is
for e-mail reminders; we can't short-circuit this logic.

This matches how the web client works as well, at least from a couple of
tests I did.