fix*: rsvp: Don't report progress for recurring events
Figuring out whether a recurring event has an instance that happens to
be ongoing is somewhat awkward and heavy (cf. RecurIterator), so let's
simply report all recurring events as pending.

This is slightly invalid - doesn't take into account the `UNTIL` rule,
to name a thing - but it's good enough.

This is also how Proton Mail's RSVP widget works at the moment, on web
at least.