fix: Present a clear fork API
When forking a session, we need to pass in something like `ios-mail`.
However, the existing `fork` method in `proton-core-common` was
incorrectly passing in the full app version (eg `ios-mail@7.0.1`).
Luckily, that was unused, so I deleted it.

The other public methods `fork_with_version` accepted a single parameter
which made it easy to make mistakes. I renamed these methods to `fork`
to replace the method I removed, and I changed the signature to accept
separate `platform` and `product` strings to prevent misuse.