fix(ET-2992): Open PGP attachments via `get_attachment`
PGP attachments can not be synced from the server and do not have a
remote id as they are embedded in the encrypted body.

The clients use `Attachment::get_attachment` to retrieve the
attachments. This method tries to sync the attachment if it is
incomplete. The method has been updated to skip PGP attachments since
we have all the required data locally already.