fix: Share extension is loading user key secret
We are using TempStore in the share extension to prevent ShareExt
from writing to the Database when acquiring a session from the API.

It is to prevent a race condition where two processes (share ext and
main app) are for example getting the session in the same time.

TempStore does not have an access to the DB at all, its in-memory.

The downside is, if you forget to populate all its necessary fields,
it will return `None` always, causing `PGPKeyAccess(NoUserSecret)`
error when creating/saving/sending a draft.