feat: Action Auto dependencies
Update all actions to have dependencies for parallel execution in the
default group.

Setup occurs via the `ActionDependencyKeysBuilder` helper which
provides some utility functions to make this easier to use compared to
the original proposal a while back.

Certain actions also have internal grouping. E.g: read/unread actions
have a group key so that if a successive read/unread chain can be
properly cancelled.

Finally, there is still one small change left before we can enable the
parallel execution. We need to record all ids for a given dependency key
so that the event loop can work correctly. Currently it will only wait
for the last action that depends on the event loop, but this does not
guarantee it will be the last action to be executed. To be address as a
follow up.