refactor*: Record all dependency keys in action-queue
Rather than only recording the last action id, record all action ids
for the key.

The previous approach could result in cases where a certain chain of
optional actions could bypass previous action if for some reason the
user cancelled the last action with that key, but we still had other
actions in the queue dependent on the same key.

This crucial bit is required to ensure the event loop is scheduled
correctly.