fix(ET-4124): Fetch event in "groups"
If you have a very active user with a lot of separate events, seeing
the changes of those events can take a while if we fetch all 50 of them
in one go.

The code has been updated to fetch events that are "grouped" together.
See `collect_events` for more details.

Test have also been ported to validate the new behavior.