feat: Expose context builder when building context in mail app
Mail context does not have own Context registry (and that is a good thing)
However, we are still interested in registering services that are
relevant only to our mail application. By exposing `ContextBuilder` in
`Context::new` we can achieve that for free.

# Additional cleanup

Patryk has noticed that `as_any()` trick is not needed if you dereference
pointer in the right way. Simplifies our code.