fix: When scroller finds out on refresh that the label has items to display but they are not currently retivable it will queue `FetchMore` task
So this is very tricky bug I have found. Basically we cannot fetch more and push update to the client before running diff because we will find out soon enough that the user of the scroller gets incorrect data displayed (with possibility of missing parts of the newest data) when having empty scroller and non-empty update.