feat: Network Monitor Service
Replacement for the current status observer and watcher. The current
implementations are not centralized and can be influenced by many
different sessions.

This version centralizes the network checking via a dedicated network
checker. This also makes it easier to integrate OS level feedback about
when something changed.

We also have an immediate check mode that does not wait for the
background check, but it is rate limited to avoid excessive requests.

After the merge, the current `StatusObserver` and `StatusWatcher` will
be removed and our `Session` types will be initialized with a
`ConnectionMonitor` or use `Standalone` for fallback where this is
currently not possible.