fix(ET-4697): Malformed declarations do not affect valid CSS declarations
# What's broken.

We are using lightningCSS which is using Servo rust-cssparser crate.
That crate when encountering a missing token returns EOF error.

However, we are using `error_recovery: true` option. It should just try to recover
and continue parsing stylesheet.

The problem is, this recovery is broken - it discards **everything** including proper
valid CSS declarations UNTIL the broken declaration.

Example:

This is a broken CSS:

```
div {
  color: red;
  important;
  background-color: blue;
}
```

Note, how `important;` has no `:`. CSS parser expects there to be a `key:value;` pair,
so it returns EOF error at `;` token. It expected `:` instead. However because of
broken error recovery it drops everything including "color: red". Result:

```
div {
  background-color: blue;
}
```

Whoops. Suddenly our color is no longer red. That is problematic for Apple,
because in that incorrectly trimmed content was a declaration
`display: block !important;` that was overriding previous `display: none;`!

In other words, because of this bug we render an empty email.

# Solution

Since we are maintaining Lightningcss fork, we have made a new commit that
fixes the issue directly in the parser.

Link to the commit: https://github.com/frondeus/lightningcss/commit/ab7380af1465412bcaeb403d52cfa108165346e0

## Additional work in this commit

Additionally, to ensure we always use error recovery, I have extracted
stylesheet and style attribute parsing to separate functions