fix(ET-4170): Convert newlines in mobile signatures
Mobile signature is an HTML payload, which means - in particular - that
it doesn't care about newlines; if you enter:

```
Cheers,
Patryk
```

... and then send an HTML email, what you'll see in it is:

```
Cheers, Patryk
```

This seems to confuse users in practice, so this commit changes the
logic to transparently convert newlines into `<br />` tags.

The previous version of Proton Mail already does this (at least on iOS,
didn't cross-check with Android), so this is not really a new feature,
more of a compatibility thing.

Note that we don't convert the `br` tags into newlines when user views
the mobile signature back again - that is, if you save:

```
Cheers,
Patryk
```

... and then re-open the settings, what you'll see in the signature
composer is:

```
Cheers,<br />Patryk
```

... this is expected and that's how the previous app also works.