fix: Event Loop Locking
The current implementation can run into a problem where if you have bad
connection and the user does a pull to refresh, the queue can stall due
to event poll subscriber lock.

Eventually the problem goes away, but due to network issues this can
take a while. It's also possible the lock may remain locked if the queue
task gets paused or aborted.

Additionally, this was also causing a lot of parallel event poll requests.

To counter this, we move the event poll to an actor to avoid any locks
to begin with, and this way we also guarantee there is only one source
that ever performs the event poll.