refactor: Stash connection acquire timeouts
We have observed some issues on android where in certain circumstances
we can generate a db locked state if during transaction we attempt to
acquire a new connection and none are a available. The most common of
these cases is the account db access for the user secrets.

We now enforce that all connections need to be acquired in 15 seconds or
we will bail with an error. This will at least unblock the app with an
error at the call site rather than forcing the user to restart the app.