fix(ET-4384): Revert default action queue to 1 worker
Clients started to report some mismatch between order of updates when
mixing operations on conversations and messages.

Since prefetching now runs in its own queue the impact of this should
not be severe and order of operations is preserved.

This also reduces the chance of the event loop running at the same time
as another action.