fix(ET-4867): Update move action to respect keep messages setting
There is a mail setting that can be enabled which keeps the original
draft/sent message in the same location after a move operation. We did
not respect this setting.

The move_to code has been updated to check the mail settings for this
value and if we have the `ShowMoved::DoNotKeep` value we optimistically
remove the sent or drafts label to match the expected server state.