refactor: Delay event poll until all actions have executed
We are currently cancelling the event poll when actions are queued, but
since the actions can take a while to execute, the event loop action
will get queued during action execution.

We have observed that if this executed in quick succession after a large
number of actions, it can bring back old state temporarily until the
next event poll.

We no cancel and delay the event poll when the user queues an action and
when an action finishes executing to give the server more time to
process the data and to avoid out of date states.