fix: Temporary - prevent getPrevious from showing the same item if the item was removed
This is a temporary non ideal but fast solution.

# The problem

Imagine a list of [A, B, C, D]. We are currently displaying "B".
Our siblings are "A" and "C".
We are moving "B" to trash - it had position 1.

the next sibling is "C"/2.
We go forward, so we try to find "C". Because our list is mutated in
the separate thread, there is a risk that it still contains "B".
Let's assume that scenario.

In that case "C" is still under position 2.
We try to find new siblings, of "C". Because list did not udpate yet,
we get "B"/1 and "D"/3.

But after we updated siblings pointers, we release mutex lock, and
the list is free to be updated. The result is:
[A, C, D] with pointers "B"/1 and "D"/3.

Now if we ask for previous sibling, we search for "B". We fail.
And our fallback tells us to use its position which is 1. But now,
under the 1 lies "C". So "previous element of C is C".

The problem is, when we update the list we do not invalidate/update
pointers to this list.

# Right solution

The corrent way would be for example by tracking all pointers and
updating them when list is updated and items are removed.
However this is complex and we already are writing untested code in
the wrong layer (uniffi).

Tracking issue: https://protonag.atlassian.net/browse/ET-5057

# "Workaround" temporary solution.

We could assume that for the auto-swipe operation only one item is
removed at the time. That means, if we search for "B" and we fail,
we should not ask for 1, but ask for 1-1 = 0. Because we know that
(old_position_of_previous - number_of_elements_removed) works if we
assume that number_of_elements_removed is always 1.

After moving to the mail-common layer we could return to the pointer
tracking solution and have it heavily tested.