feat(ET-5097): Scroller fetch new feadback
Provide feedback to the UI via new scroller updates when the scroller is
fetching new data from the server.

To make it easier for the clients to process these updates the
`ScrollerUpdate` has been split into sub categories. There is a sub
category for list operations, status updates and general purpose error
updates.