feat(ET-5183): Rebase local state changes
Introduce the `rebase_local` method to the action handler which
will be invoked when we want to rebase the action state when we pull in
new state from the server that undoes the local state.

In this first version we use a brute-force approach and simply re-apply
all the queued changes. All the actions _should_ perform idempotent
operations and on, average, the selected number of item should be small.

Every time the event loop runs or we fetch new data from the server we
reapply the actions to ensure local changes are visible on top of the
new state.

There is still a conflict of state such as when you mark a conversation
as read locally and then a new message arrives for the conversation, the
full conversation will be marked as read again. This will be handled
in a follow up as it requires more extensive changes and we also need
API support (INBE-2510) so the server applies that exact change.

The current setup should be "good enough" got gather early feedback on
this approach while we work on optimizing the number of required actions
to run when rebasing.