fix(ET-5389): attachment-upload: Keep action alive during apply_remote()
This commit extends `apply_remote()` for `AttachmentUpload` so that
every 10s it performs a touch that refreshes action permit's timestamp.

---

Each action has an implied timeout of 60s:

```rust
pub(crate) const DEFAULT_LOCK_TIMEOUT: Duration = Duration::from_secs(60);
```

... after which it's "up for grabs" again, reasoning being - the action
must've gotten stuck / must've crashed, let's try executing it again.

Of course, this does not play with actions that can simply take a long
time to execute, such as uploading a large attachment on slow network -
in cases like these it's possible we'd try to upload the same attachment
twice, since we can't distinguish between "action is dead" and "action
is running, albeit slowly".

(something something the halting problem)

---

Note that I think this is a hacky solution, because in principle all
`apply_remote()`s can be slow and there's no reason to bless only the
attachment upload action; I think the entire "action is up for grabs
again" mechanism needs to be re-thought, but parallel to this small fix.