feat: Legacy feature flags - allow Proton to override overriden values
# Example of current, broken status quo

I will flag correct state with [v] - correct or [x] - incorrect

1. flag "X" is set to false
2. API returns false during a refresh
    a) we store enabled=false, overriden=NULL
    b) SDK returns false   [v]
3. User executes action to override to true
    a) we store enabled=false, overriden=true
    b) SDK returns true    [v]
4. Action queue executes remote side of the action
    a) Backend stores new value.
5. Refresh is triggered because of <timeout/foreground event>
    a) API returns enabled=true. There is no way to know if its overriden
        value or new value set by Proton employee.
    b) we store enabled=true, overriden=true
    c) SDK returns true    [v]
6. A year passes, now Proton wants to set this flag again to `false` for
    this particular user! Currently it's impossible because:
    a) API returns enabled=false.
    b) we store enabled=false BUT overriden stays =true
    c) SDK returns true    [x] <--- This is incorrect.

I incorrectly assumed that overriden value can be only changed by the
user. This is not what we want for rating booster.

# Solution

1. When we send PUT request, the API responds with new feature flag state

We can use that to update the flag immediately. Moreover we can store
`overriden_at` in new column so that we can avoid having race conditions.

2. overriden_at ALWAYS comes from backend. It means there is only **one**
source of the clock, so we are safe to compare overriden_at with response
updated_at values. That's how we can detect that during the refresh
the data is stale!

Timeline of this potential data race:

1. First user updates the override
2. Refresh background tasks STARTS asking for new flags but our backend
has a gigantic hiccup or there is a network connection latency.
3. Meanwhile apply_remote is triggered, backend now has new value after
successful PUT (latency can be asynchronous).
4. Finally background task gets legacy feature flags with stale data.

By comparing overriden_at and updated_at we can avoid it.

Additionally: If overriden_to exists but overriden_at does not, it means
that the overriden_to was not yet processed by the backend.
But as soon as it is processed, we can safely remove
overriden_to and overriden_at but during the next REFRESH.