feat(ET-5182): Core and Contact v6 event loops
Setup core and contacts v6 event loop handlers. We try to share as much
code with the v5 event loop as possible so both can be running depending
on the build feature flag.

In v5 mode we also start recording the event ids for the contacts and
mail loops ahead of time, so we can switch between the two. A migration
was also added to pre-initialize this data.