fix(ET-5586): Fetch label counters with scroller fetch new
When fetching new data with the scroller, also refetch the label counter
for the current label so that is up to date with the new
conversations/messages being pulled in.

Unfortunately, the save of this data happens in a untracked transaction.
Enabling tracking transactions seems to alter the behavior of the
scroller in sublte ways.

To prevent odd behaviors from creeping in, we issue a fake update to
trigger the watchers and ET-5589 to revisit this when we have more time.